Split module into header and src files
